backend/requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2-binary
djangorestframework-simplejwt
pgvector
openai
python-dotenv
sentence-transformers
django-q2
gunicorn
whitenoise
dj-database-url
